Refactoring ProductCreate and ProductEdit, moving Controllers to custom components
